ENH: Drop the abandoned markup snap-on-place path
The vessel annotation no longer routes through Slicer markups (ADR-0037):
placement + snap live on the LayerDM TerritoryPlacementPipeline against
the annotation carrier.  Remove the exploratory markup snap-on-place that
predated that decision -- snap_control_point_to_surface + its export, and
the widget's onEndpointPlaced PointPositionDefinedEvent observer + its
_snapping/_observedEndpointsNode state.  updateHighlightWiring keeps only
the surviving hover pickSurface/visibility sync (transitional scaffolding
Stage 2 retires with the markups selector).  closed_surface_polydata (the
shared surface seam the hover + placement pipelines both use) stays.  The
snap tests are dropped; the surviving widget-wiring pin moves to
test_vessel_highlight_wiring.py.
