+++ VascularTerritories/VascularTerritoriesLib/__init__.py
@@ -12,7 +12,6 @@ from __future__ import annotations
 from .VesselSurfacePick import VesselSurfacePick
 from .VesselHighlightWiring import (
     closed_surface_polydata,
-    snap_control_point_to_surface,
 )
 
 # The LayerDM highlight Pipeline import depends on LayerDMLib (reachable
@@ -43,7 +42,6 @@ except ImportError:  # pragma: no cover - LayerDMLib unreachable bare
 __all__ = [
     "VesselSurfacePick",
     "closed_surface_polydata",
-    "snap_control_point_to_surface",
     "VesselHighlightPipeline",
     "registerVesselHighlightPipelineCreator",
     "TerritoryPlacementPipeline",
